add some screenreader-only text to the top of the part feedback box.
The contents of the part feedback box are read out when they change.
When you've moved focus to another part, you have no context to work out
which part the announced feedback is for.

This adds a line "Feedback for <part name>." to the top of the feedback
box, to give some context.