working towards dark mode and customisable colours
I've ended up removing huge swathes of Bootstrap, with the aim of
reducing the number of colours I have to think about, and to simplify
the stylesheet.

I've copied in the CSS variables from the LTI provider and tried to
apply them.
I think that's complete with spacing, but colours are proving very
tricky.

The next thing to do is to tackle dark mode. Maybe users will be able to
choose the light or dark theme and just choose text, background and
maybe link colours.

There are lots and lots of shades of grey that should be merged.

I made https://which-colours-are-similar.think.somethingorotherwhatever.com/
to visualise the full set of colours.