all the changed feedback settings have different names
We realised that old version of the LTI provider wouldn't be able to
interpret the feedback settings whose type has changed, for example
'showactualmarks' changed from a boolean to a string.

So now all of the changed feedback settings have different names, and
the editor will keep producing the old settings with values roughly
equivalent to the new settings.