Finish removing bootstrap (and jQuery?)
All of the bootstrap JS and CSS has gone.

The loading bar is now quite small and in the middle of the screen.
Having it occupy the entire viewport felt quite bad for
photosensitivity.

Numbas.display.init now just toggles a class 'loaded' on the <body> tag,
instead of showing/hiding a load of elements. CSS makes sure that the
`#loading` element disappers and the main display appears.

The `active` class is gone, using the `aria-current` attribute instead.

The `col-*`, `container-*`, `form-control` and `form-group` classes are gone.

All input and select elements get the same styling as answer inputs used
to. This affects the name and password inputs on the front page, and the
inputs on the display options modal.

The display code doesn't use jQuery anywhere.

The jme scope information used by MathJax is just set on the element
object, instead of using jQuery data.

The numberentry display 'showPreview' and 'studentAnswerLaTeX'
observables are gone.

The old 'question-html-attached' event is no longer triggered. Instead,
use the 'HTMLAttached' signal of the question object.

I've removed unused knockout binding handlers.

Still need to look at the menu and diagnostic modes.